requirements.txt
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
Flask-Caching==2.0.2
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
oracledb==1.3.1
packaging==23.1
pandas==1.3.5
pip==22.0.4
plotly==5.14.1
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
setuptools==67.7.2
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.2.2
typing_extensions==4.5.0
Werkzeug==2.2.3
wheel==0.40.0
zipp==3.15.0
